LU-5417 obdclass: fix comparison between signed and unsigned
Make lu_buf->lb_len unsigned.

+++ lustre/osp/osp_object.c
@@ -504,7 +504,7 @@ static int osp_xattr_get_interpterer(const struct lu_env *env,
 	LASSERT(ooa != NULL);
 
 	if (rc == 0) {
-		int len = sizeof(*oxe) + oxe->oxe_namelen + 1;
+		size_t len = sizeof(*oxe) + oxe->oxe_namelen + 1;
 
 		rc = object_update_result_data_get(reply, rbuf, index);
 		if (rc < 0 || rbuf->lb_len > (oxe->oxe_buflen - len)) {
